+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.3.1
 - Add `tracciatto.forceEnableRdbgDebugType` setting to force-enable `rdbg` debug type support even when `vscode‑rdbg` is installed but detected as inactive.
+- **Attach to…** now offers autodected `rdbg` sockets.
 
 ## 0.3.0
 - Add extensible Ruby version manager support, including built‑in integration for `asdf`, `rbenv`, and `rvm`, plus a fully customizable `custom` mode for any environment tool.